add basic layout and responsive stylings
