--- src/assets/styles/grid.scss
@@ -1,4 +1,0 @@
-$grid-layout-gap: 20px;
-
-$headerHeight: 5em;
-$footerHeight: 5em;
