+++ src/assets/styles/main.scss
@@ -1,5 +1,5 @@
-@import './vendor/normalize';
-@import 'fonts';
+@import "./vendor/normalize";
+@import "fonts";
 
 html {
   height: 100%;
