+++ src/assets/styles/variables.scss
@@ -1,0 +1,5 @@
+$headerHeight: 4em;
+$footerHeight: 16em;
+
+$base-padding: 1em;
+$base-margin: $base-padding;
