--- src/common/layout/pageLayout.module.scss
@@ -1,36 +1,0 @@
-@import 'layout';
-@import 'grid';
-
-
-.pageHeader {
-    grid-area: pageHeader;
-    background-color: red;
-}
-
-.pageBody {
-    grid-area: pageBody;
-    background-color: green;
-}
-
-.pageFooter {
-    grid-area: pageFooter;
-    background-color: yellow;
-}
-
-.pageWrapper {
-    height: 100%;
-    width: 100%;
-
-    display: grid;
-    grid-template-areas: "pageHeader"
-                         "pageBody"
-                         "pageFooter";
-    grid-template-rows: $headerHeight 1fr $footerHeight;
-    grid-template-columns: repeat(12, [col-start] 1fr);
-}
-
-@include respond-above(lg) {
-    .pageBody {
-        grid-column: col-start / span 4fr;
-    }
-}
