+++ src/domain/app/footer/footer.module.scss
@@ -1,0 +1,36 @@
+@import "colors";
+@import "mixins";
+@import "variables";
+@import "fonts";
+
+$logoHeight: 4.375em;
+$logoWidth: 9.5em;
+
+.footer {
+    display: grid;
+}
+
+.footerWrapper {
+    @include koro('basic', $hel-light, 1000);
+    
+    display: grid;
+    align-items: center;
+}
+
+.helsinkiLogo {
+    @include helsinki-logo($hel-black);
+    background-position: center;
+    height: $logoHeight;
+    width: $logoWidth;
+    margin-bottom: $base-margin;
+}
+
+.copyright{
+    border-top: 1px solid $hel-black;
+    padding-top: $base-padding;
+
+    p {
+        margin: 0;
+        color: $hel-gray;
+    }
+}
