+++ src/domain/app/header/Header.tsx
@@ -2,14 +2,19 @@ import React from 'react';
 
 import styles from './header.module.scss';
 import { formatMessage } from '../../../common/translation/utils';
+import Container from '../layout/Container';
 
 type Props = {};
 
 function Header(props: Props) {
   return (
-    <header className={styles.header}>
-      <span className={styles.logo}></span>
-      <h1 className={styles.appName}>{formatMessage('appName')}</h1>
+    <header className={styles.headerWrapper}>
+      <Container>
+        <div className={styles.navbarTop}>
+          <div className={styles.logo}></div>
+          <h3 className={styles.appName}>{formatMessage('appName')}</h3>
+        </div>
+      </Container>
     </header>
   );
 }
