+++ src/domain/app/header/header.module.scss
@@ -1,14 +1,19 @@
-@import 'mixins';
-@import 'colors';
+@import "mixins";
+@import "colors";
+@import "variables";
 
-$logo-width: 5em;
-$header-height: 5em;
+$logoHeight: 2.5em;
+$logoWidth: 5.5em;
 
-.header {
-  align-items: center;
+.headerWrapper {
+  display: grid;
+}
+
+.navbarTop {
+  height: 100%;
   display: flex;
-  padding: 0 10px;
-  height: $header-height;
+  flex-direction: row;
+  align-items: center;
 }
 
 .appName {
@@ -16,8 +21,8 @@ $header-height: 5em;
 }
 
 .logo {
-  @include helsinki-logo(black);
+  @include helsinki-logo($hel-black);
   background-position: center;
-  width: $logo-width;
-  height: inherit;
+  height: $logoHeight;
+  width: $logoWidth;
 }
