+++ src/domain/app/layout/Container.tsx
@@ -1,0 +1,9 @@
+import React, { FunctionComponent } from 'react';
+
+import styles from './container.module.scss';
+
+const Container: FunctionComponent = ({ children }) => {
+  return <div className={styles.container}>{children}</div>;
+};
+
+export default Container;
