+++ src/domain/app/layout/container.module.scss
@@ -1,0 +1,22 @@
+@import "layout";
+
+.container {
+    display: grid;
+    grid-row: span;
+
+    & > * {
+        grid-column: 2;
+    }
+}
+
+@include respond-above(sm) {
+    .container {
+        grid-template-columns: 1fr 10fr 1fr;
+    }
+}
+
+@include respond-below(sm) {
+    .container {
+        grid-template-columns: 5% 1fr 5%;
+    }
+}
